# api/openapi.yaml
openapi: 3.0.3
info:
  title: Price Hunter API
  version: "1.0"
  description: |
    Track product prices and get alerted when they reach a target.

    * Money is `{"amount": "94.00", "currency": "USD"}`, a decimal string, never a float.
    * Errors use RFC 9457 `application/problem+json`.
    * Lists are paginated with an opaque `cursor`.
    * Products owned by other users return 404, never 403.
servers:
  - url: http://localhost:8088
    description: Local server mode
security:
  - bearer: []
paths:
...
  /v1/products:
    post:
      summary: Track a product
      operationId: createProduct
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: "#/components/schemas/CreateProductRequest" }
            example:
              name: Ninja Air Fryer
              url: https://shop.example.com/p/ninja-af101?utm_source=newsletter
              target_price: { amount: "100.00", currency: USD }
              check_frequency: 6h
      responses:
        default: { $ref: "#/components/responses/Problem" }
        "201":
          description: Created; the first check is queued.
          headers:
            Location: { schema: { type: string } }
            ETag: { schema: { type: string } }
          content:
            application/json:
              schema: { $ref: "#/components/schemas/Product" }
        "400": { $ref: "#/components/responses/Problem" }
        "401": { $ref: "#/components/responses/Problem" }
        "403": { $ref: "#/components/responses/Problem" }
        "409": { $ref: "#/components/responses/Problem" }
        "413": { $ref: "#/components/responses/Problem" }
        "415": { $ref: "#/components/responses/Problem" }
        "422": { $ref: "#/components/responses/Problem" }
    get:
      summary: List the watchlist
      operationId: listProducts
      parameters:
        - { name: limit, in: query, schema: { type: integer, minimum: 1, maximum: 100 } }
        - { name: cursor, in: query, schema: { type: string } }
      responses:
        default: { $ref: "#/components/responses/Problem" }
        "200":
          description: OK
          content:
            application/json:
              schema:
                type: object
                required: [items]
                properties:
                  items: { type: array, items: { $ref: "#/components/schemas/Product" } }
                  next_cursor: { type: string }
        "400": { $ref: "#/components/responses/Problem" }
        "401": { $ref: "#/components/responses/Problem" }
  /v1/products/{id}:
    parameters:
      - { name: id, in: path, required: true, schema: { type: string } }
    get:
      summary: Get a product
      operationId: getProduct
      responses:
        default: { $ref: "#/components/responses/Problem" }
        "200":
          description: OK
          headers:
            ETag: { schema: { type: string } }
          content:
            application/json:
              schema: { $ref: "#/components/schemas/Product" }
        "404": { $ref: "#/components/responses/Problem" }
    patch:
      summary: Update name, target, frequency or paused state
      operationId: updateProduct
      parameters:
        - name: If-Match
          in: header
          description: ETag from a previous response; returns 412 if the product changed.
          schema: { type: string }
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: "#/components/schemas/UpdateProductRequest" }
            example:
              target_price: { amount: "90.00", currency: USD }
              check_frequency: 12h
      responses:
        default: { $ref: "#/components/responses/Problem" }
        "200":
          description: Updated
          content:
            application/json:
              schema: { $ref: "#/components/schemas/Product" }
        "400": { $ref: "#/components/responses/Problem" }
        "404": { $ref: "#/components/responses/Problem" }
        "412": { $ref: "#/components/responses/Problem" }
        "422": { $ref: "#/components/responses/Problem" }
    delete:
      summary: Stop tracking (history is purged asynchronously)
      operationId: deleteProduct
      responses:
        default: { $ref: "#/components/responses/Problem" }
        "204": { description: Deleted }
        "404": { $ref: "#/components/responses/Problem" }
...
components:
  securitySchemes:
    bearer:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: Cognito access/ID token, validated by the API Gateway JWT authorizer.
  responses:
    Problem:
      description: Error
      content:
        application/problem+json:
          schema: { $ref: "#/components/schemas/Problem" }
  schemas:
    Money:
      type: object
      required: [amount, currency]
      properties:
        amount: { type: string, pattern: '^\d+(\.\d+)?$', example: "94.00" }
        currency: { type: string, pattern: '^[A-Z]{3}$', example: USD }
    CreateProductRequest:
      type: object
      additionalProperties: false
      required: [name, url, target_price]
      properties:
        name: { type: string, maxLength: 120 }
        url: { type: string, maxLength: 2048 }
        target_price: { $ref: "#/components/schemas/Money" }
        check_frequency: { type: string, enum: ["1h", "6h", "12h", "24h"], default: "6h" }
    UpdateProductRequest:
      type: object
      additionalProperties: false
      properties:
        name: { type: string, maxLength: 120 }
        target_price: { $ref: "#/components/schemas/Money" }
        check_frequency: { type: string, enum: ["1h", "6h", "12h", "24h"] }
        paused: { type: boolean }
    Product:
      type: object
      required: [id, name, url, retailer, target_price, current_price, lowest_price, availability,
        check_frequency, status, stale, alert, last_checked_at, last_success_at, next_check_at,
        last_error, created_at, updated_at, version]
      properties:
        id: { type: string }
        name: { type: string }
        url: { type: string }
        retailer: { type: string, description: 'Retailer profile ID, or "generic"' }
        target_price: { $ref: "#/components/schemas/Money" }
        current_price:
          allOf: [{ $ref: "#/components/schemas/Money" }]
          nullable: true
        lowest_price:
          allOf: [{ $ref: "#/components/schemas/Money" }]
          nullable: true
        availability: { type: string, enum: [IN_STOCK, OUT_OF_STOCK, UNKNOWN] }
        check_frequency: { type: string, enum: ["1h", "6h", "12h", "24h"] }
        status: { type: string, enum: [ACTIVE, PAUSED, NEEDS_ATTENTION, GONE] }
        stale: { type: boolean, description: Last successful price is older than 3 check intervals }
        alert:
          type: object
          required: [state, triggered_at]
          properties:
            state: { type: string, enum: [ARMED, TRIGGERED] }
            triggered_at: { type: string, format: date-time, nullable: true }
        last_checked_at: { type: string, format: date-time, nullable: true }
        last_success_at: { type: string, format: date-time, nullable: true }
        next_check_at: { type: string, format: date-time, nullable: true }
        last_error:
          type: object
          nullable: true
          required: [code, retryable]
          properties:
            code: { $ref: "#/components/schemas/OutcomeCode" }
            retryable: { type: boolean }
        latest_check:
          type: object
          required: [id, status]
          properties:
            id: { type: string }
            status: { $ref: "#/components/schemas/CheckStatus" }
        created_at: { type: string, format: date-time }
        updated_at: { type: string, format: date-time }
        version: { type: integer, description: Increments on user edits; also exposed as the ETag }
...
    CheckStatus:
      type: string
      enum: [QUEUED, RUNNING, SUCCEEDED, FAILED, SKIPPED]
    OutcomeCode:
      type: string
      enum: [OK, PRICE_NOT_FOUND, AMBIGUOUS_PRICE, IMPLAUSIBLE_PRICE, CURRENCY_MISMATCH, FETCH_TRANSIENT,
        RATE_LIMITED, BLOCKED, ROBOTS_DISALLOWED, NOT_FOUND, FORBIDDEN_TARGET, UNSUPPORTED_CONTENT,
        TOO_LARGE, HTTP_ERROR, CIRCUIT_OPEN, PRODUCT_INACTIVE]
...
    Problem:
      type: object
      required: [type, title, status]
      properties:
        type: { type: string }
        title: { type: string }
        status: { type: integer }
        detail: { type: string }
        instance: { type: string }
        request_id: { type: string }
        existing_product_id: { type: string }
        errors:
          type: array
          items:
            type: object
            required: [field, code, message]
            properties:
              field: { type: string }
              code: { type: string }
              message: { type: string }